requirements.txt
Flask==1.1.1
Flask-HTTPAuth==3.3.0
line-bot-sdk==1.14.0
numpy==1.17.2
requests==2.22.0
beautifulsoup4==4.8.1
aiohttp==3.6.1
